+++ spring-boot-extension-tests/redis-spring-boot-test/redisearch-spring-boot-test/redisearch-spring-boot-test-webmvc/src/main/java/com/livk/redisearch/mvc/RediSearchApp.java
@@ -16,7 +16,7 @@
 
 package com.livk.redisearch.mvc;
 
-import com.livk.commons.beans.BeanLambda;
+import com.livk.commons.util.BeanLambda;
 import com.livk.commons.jackson.util.JsonMapperUtils;
 import com.livk.context.redisearch.StringRediSearchTemplate;
 import com.livk.redisearch.mvc.entity.Student;
