+++ spring-extension-commons/src/main/java/com/livk/commons/util/BeanLambdaDescriptor.java
@@ -1,12 +1,9 @@
 /*
  * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
  *       https://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,11 +11,8 @@
  * limitations under the License.
  */
 
-package com.livk.commons.beans;
+package com.livk.commons.util;
 
-import com.livk.commons.util.BeanUtils;
-import com.livk.commons.util.ClassUtils;
-import com.livk.commons.util.ReflectionUtils;
 import lombok.Getter;
 import lombok.SneakyThrows;
