+++ spring-extension-commons/src/main/java/com/livk/commons/util/RecordWrapper.java
@@ -1,12 +1,9 @@
 /*
  * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
  *       https://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,33 +11,15 @@
  * limitations under the License.
  */
 
-package com.livk.commons.selector;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
+package com.livk.commons.util;
 
 /**
- * <p>
- * SpringAbstractImportSelectorTest
- * </p>
+ * 使用record进行默认实现
  *
+ * @param <V> 相关实例
+ * @see GenericWrapper#of(Object)
  * @author livk
  */
-class SpringAbstractImportSelectorTest {
-
-	@Test
-	void testFindAnnotation() {
-		MyAnnotationImportSelector selector = new MyAnnotationImportSelector();
-		assertEquals(MyAnnotation.class, selector.getAnnotationClass());
-	}
-
-	@interface MyAnnotation {
-
-	}
-
-	static class MyAnnotationImportSelector extends SpringAbstractImportSelector<MyAnnotation> {
-
-	}
+record RecordWrapper<V>(V unwrap) implements GenericWrapper<V> {
 
 }
