+++ spring-extension-commons/src/test/java/com/livk/commons/spring/SpringAbstractImportSelectorTest.java
@@ -1,9 +1,12 @@
 /*
  * Copyright 2021-2024 spring-boot-extension the original author or authors.
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
+ *
  *       https://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -11,15 +14,33 @@
  * limitations under the License.
  */
 
-package com.livk.commons.beans;
+package com.livk.commons.spring;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * 使用record进行默认实现
+ * <p>
+ * SpringAbstractImportSelectorTest
+ * </p>
  *
- * @param <V> 相关实例
- * @see com.livk.commons.beans.GenericWrapper#of(Object)
  * @author livk
  */
-record RecordWrapper<V>(V unwrap) implements GenericWrapper<V> {
+class SpringAbstractImportSelectorTest {
+
+	@Test
+	void testFindAnnotation() {
+		MyAnnotationImportSelector selector = new MyAnnotationImportSelector();
+		assertEquals(MyAnnotation.class, selector.getAnnotationClass());
+	}
+
+	@interface MyAnnotation {
+
+	}
+
+	static class MyAnnotationImportSelector extends SpringAbstractImportSelector<MyAnnotation> {
+
+	}
 
 }
