+++ spring-extension-commons/src/test/java/com/livk/commons/util/GenericWrapperTest.java
@@ -1,12 +1,9 @@
 /*
  * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
  *       https://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,7 +11,7 @@
  * limitations under the License.
  */
 
-package com.livk.commons.beans;
+package com.livk.commons.util;
 
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Test;
